Agregando migraciones y corrigiendo el place

+++ app/Http/Controllers/APIBeta/PlaceController.php
@@ -4,27 +4,43 @@ namespace App\Http\Controllers\APIBeta;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\PlaceRequest;
+use App\Image;
 use App\Place;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Str;
 
 class PlaceController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+    //constructor 
+    public function __construct()
+    {
+        $this->middleware('auth:api');
+    }
     public function index()
     {
-        $places = Place::paginate(10);
-
+        $places = auth()->user()->places()->paginate();
         return response()->json($places, 200);
     }
 
-    
+
     public function store(PlaceRequest $request)
     {
-        Place::create($request->all());
+        $place = auth()->user()->places()->create($request->all());
+        if ($request->hasFile('images')) {
+            $images = $request->file('images');
+            if (!is_array($images)) {
+                $images = [$images];
+            }
+            foreach ($images as $image) {
+                $custom_name = 'img-' . Str::uuid()->toString() . '.' . $image->getClientOriginalExtension();
+                $place->imagenes()->create([
+                    'image_path' => $custom_name,
+                ]);
+                $image->move(public_path() . '/eventos', $custom_name);
+            }
+            
+        }
         return response()->json(['status' => 'success', 'message' => 'Se Registró correctamente'], 201);
     }
 
@@ -36,8 +52,11 @@ class PlaceController extends Controller
      */
     public function show($id)
     {
-        $place = Place::findOrfail($id)->load('imagenes');
-        return response()->json($place, 200);
+        if ($this->validatePlace()) {
+            $place = Place::findOrfail($id)->load('imagenes');
+            return response()->json($place, 200);
+        }
+        return response()->json(['error' => 'No tiene los permisos para ver este recurso'], 404);
     }
 
     /**
@@ -60,8 +79,11 @@ class PlaceController extends Controller
      */
     public function update(PlaceRequest $request, Place $place)
     {
-        $place->update($request->all());
-        return response()->json(['status' => 'success', 'message' => 'Se Actualizó correctamente'], 200);
+        if ($this->validatePlace()) {
+            $place->update($request->all());
+            return response()->json(['status' => 'success', 'message' => 'Se Actualizó correctamente'], 200);
+        }
+        return response()->json(['status' => 'error', 'message' => 'No tiene los permisos necesarios'], 404);
     }
 
     /**
@@ -72,7 +94,15 @@ class PlaceController extends Controller
      */
     public function destroy(Place $place)
     {
-        $place->delete();
-        return response()->json(['status' => 'success', 'message' => 'Se Eliminó correctamente'], 200);
+        if ($this->validatePlace()) {
+            $place->delete();
+            return response()->json(['status' => 'success', 'message' => 'Se Eliminó correctamente'], 200);
+        }
+        return response()->json(['status' => 'error', 'message' => 'No tiene los permisos necesarios'], 404);
+    }
+
+    public function validateAction($user_id)
+    {
+        return $user_id === auth()->user()->id || auth()->user()->id_rol === 1;
     }
 }
